requirements.txt
Flask>=2.0
google-api-python-client>=2.70.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
APScheduler>=3.9
pytrends>=4.8.0
requests>=2.28